+++ src/FP/ch/wozza.c
@@ -207,7 +207,7 @@ void chWozza_update(Actor *this){
             this->position[0] = local->unkC[0];\
             this->position[1] = local->unkC[1];\
             this->position[2] = local->unkC[2];
-            __spawnQueue_add_1((GenFunction_1)chWozza_spawnJiggy, reinterpret_cast(s32, this->marker));
+            __spawnQueue_add_1((GenFunction_1)chWozza_spawnJiggy, reinterpret_cast(intptr_t, this->marker));
             local->unk30 = FALSE;
         }
     }//L8038F910
