+++ src/SM/ch/vegetables.c
@@ -218,7 +218,7 @@ static void __chVegetables_vegetableDeathParticles(Actor* this) {
     if (this->unk38_31) {
         this->position_y += 100.0f;
         func_802CA1CC(HONEYCOMB_17_SM_COLLIWOBBLE);
-        __spawnQueue_add_4((GenFunction_4) spawnQueue_bundle_f32, BUNDLE_1F_SM_EMPTY_HONEYCOMB, reinterpret_cast(s32, this->position_x), reinterpret_cast(s32, this->position_y), reinterpret_cast(s32, this->position_z));
+        __spawnQueue_add_4((GenFunction_4) spawnQueue_bundle_f32, BUNDLE_1F_SM_EMPTY_HONEYCOMB, reinterpret_cast(intptr_t, this->position_x), reinterpret_cast(intptr_t, this->position_y), reinterpret_cast(intptr_t, this->position_z));
     }
 
     timed_mapSpecificFlags_setTrue(1.5f, SM_SPECIFIC_FLAG_7);
