+++ src/TTC/ch/treasurehunt.c
@@ -157,7 +157,7 @@ void chTreasurehunt_checkStepProgress5(NodeProp *this, ActorMarker *arg1){
         particleTargetPosition[1] = (f32)this->position_y;
         particleTargetPosition[2] = (f32)this->position_z;
         particleTargetPosition[1] =  mapModel_getFloorY(particleTargetPosition);
-        __spawnQueue_add_4((GenFunction_4)spawnQueue_actor_f32, 0xF4, reinterpret_cast(s32, particleTargetPosition[0]), reinterpret_cast(s32, particleTargetPosition[1]), reinterpret_cast(s32, particleTargetPosition[2]));
+        __spawnQueue_add_4((GenFunction_4)spawnQueue_actor_f32, 0xF4, reinterpret_cast(intptr_t, particleTargetPosition[0]), reinterpret_cast(intptr_t, particleTargetPosition[1]), reinterpret_cast(intptr_t, particleTargetPosition[2]));
         particleEmitter = partEmitMgr_newEmitter(3);
         particleEmitter_setRGB(particleEmitter, sChTreasurehuntParticleRGB);
         particleEmitter_setSprite(particleEmitter, ASSET_700_SPRITE_DUST);
