+++ src/core2/ch/icecube.c
@@ -241,7 +241,7 @@ void __chicecube_die(ActorMarker *marker, ActorMarker *other_marker){
     func_8035A04C(actor->position, 12, ASSET_505_MODEL_ICECUBE_CHUNK, actor->scale);
     func_8035A228(actor->position, 6, ASSET_700_SPRITE_DUST, actor->scale);
     if(actor->actorTypeSpecificField != ICECUBE_SPECIFIC_FIELD_1_LITTLE_CUBES){
-        __spawnQueue_add_1((GenFunction_1)__chicecube_spawnHalfCubes, reinterpret_cast(s32, actor->marker));
+        __spawnQueue_add_1((GenFunction_1)__chicecube_spawnHalfCubes, reinterpret_cast(intptr_t, actor->marker));
     }
     marker_despawn(actor->marker);
 }
